Add getLogger() information (Untested)
Which is for onDisable - So if the FormAPI plugin cannot begin to install, or if the plugin's not found within the server, it'll disable along with the message of why it's being disabled. That way, it's more easier to cope with the reasoning of why you can't enable this plugin, and makes it less confusing.
And for if the FormAPI plugin has been found, the Logger should be able to notify you regarding this as well, just cuz why not? It makes things less confusing, and knowing what to do if a plugin cannot be found, or can be found. Also, the reasoning on the Enable state as well.
This build is currently untested, so please test this before making an issue.