Add onLoad() functions (Untested)
This basically allows us to properly load plugins before the API plugin (this plugin) is fully enabled. We try this so that onLoad() can allow anyone to retry with the actual enabling state, so say if one plugin tries to enable, which is a practice Loading plugins, then the real enabling state happens, which may be harder to enable its self if you don't have the plugins required for this API system.
This is currently untested, so if there are bugs, please let me know. Thank you.